Refactored tests. Placed DummyTransducer in utils file
